+++ CHANGELOG.md
@@ -41,6 +41,7 @@
 - Improved general testing infrastructure
 - Added test cases for the index function in `Controllers/Course`
 - Added test cases for the depts function in `Controllers/Course`
+- Removed unused components `Search` and `Post`, as well as backend routes/functions used exclusively by these components
 - Added test cases for the courseInfo function in `Controllers/Course`
 - Refactored the class components in `/grid` folder to function components
 - Update stylelint integration with development environment and fix existing stylelint violations
@@ -66,6 +67,7 @@
 - Switched connection for timetable parsing from `RequireEMS` to `AllowEMS`
 - Fixed dependency caching in CircleCI configuration
 - Added dependency groups to Dependabot configuration
+- Added tests for the index function in `Controllers/Graph`
 
 ## [0.6.0] - 2024-06-24